refactor: Implement Ipfs::find_peer_info and autonat counter
This function will allow one to find a peer information from Kad. PeerInfo will be fetched if it is stored locally otherwise it will check the DHT for it.

Autonat counter is to limit the amount of autonat servers we add as it seems like the servers are not released automatically if there is any form of connection issues.

TODO: Implement a internal check to remove autonat servers there is a connection issue or is not being used. It may very well be that we only would need to add one.
(cherry picked from commit b36ba65bcf8966a1326e374a325391a65a1773d6)